+++ test/main.js
@@ -72,3 +72,6 @@ describe('Test Kapp:', () => {
   apitok(request, user, pack);
   apiuser(request, user, dbi);
 });
+
+
+// -- oOo ---